% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceCor.R
\name{VarianceCor}
\alias{VarianceCor}
\title{Variance Correlation between X (independent) and Y (dependent) variables}
\usage{
VarianceCor(x_dep, y_indep)
}
\arguments{
\item{X}{The independent variable}

\item{Y}{The dependent variable}
}
\value{
The amount of variance in Y which can be explained by X
}
\description{
Variance Correlation between X (independent) and Y (dependent) variables
}
\note{
This is *not* a causal relationship, just a statement of variance correlation
}
